ClothingCard moved to folder container
